Moving routines to uni computer
errors2.py: Minor change to ensure templatesToUse values become string
when stellar and emission line templates are combined into a single
list.

save_to_fits.py: Change to add kin_opt and pop_opt keywords to allow for
using directories other than kin and pop.

save_to_fits2.py: Started attempt to combine pickle and save_to_fits
into a single routine... Might be a lot of work for not much gain...

 On branch master
 Changes to be committed:
	modified:   errors2.py
	modified:   save_to_fits.py
	new file:   save_to_fits2.py

+++ errors2.py
@@ -1032,7 +1032,8 @@ class run_ppxf(ppxf):
 				self.templates = self.stellar_templates.templates
 			self.component = [0]*len(self.stellar_templates.templatesToUse
 				) + self.e_templates.component
-			self.templatesToUse = np.append(self.stellar_templates.templatesToUse.astype(str), 
+			self.templatesToUse = np.append(
+				self.stellar_templates.templatesToUse.astype(int).astype(str), 
 				self.e_templates.templatesToUse)
 			self.element = ['stellar'] + self.e_templates.element
 		else:
